AAUTO-555 GKI fragment defconfig added for Android automotive support.

+++ arch/arm64/include/asm/preempt.h
@@ -76,7 +76,7 @@ static inline bool __preempt_count_dec_and_test(void)
 static inline bool should_resched(int preempt_offset)
 {
 	u64 pc = READ_ONCE(current_thread_info()->preempt_count);
-	return pc == preempt_offset;
+	return pc == (u64)preempt_offset;
 }
 
 #ifdef CONFIG_PREEMPTION
