+++ include/linux/byteorder/generic.h
@@ -192,7 +192,7 @@ static inline void cpu_to_be32_array(__be32 *dst, const u32 *src, size_t len)
 {
 	int i;
 
-	for (i = 0; i < len; i++)
+	for (i = 0; i < (int)len; i++)
 		dst[i] = cpu_to_be32(src[i]);
 }
 
@@ -200,7 +200,7 @@ static inline void be32_to_cpu_array(u32 *dst, const __be32 *src, size_t len)
 {
 	int i;
 
-	for (i = 0; i < len; i++)
+	for (i = 0; i < (int)len; i++)
 		dst[i] = be32_to_cpu(src[i]);
 }
