+++ include/linux/thread_info.h
@@ -204,7 +204,7 @@ static __always_inline __must_check bool
 check_copy_size(const void *addr, size_t bytes, bool is_source)
 {
 	int sz = __compiletime_object_size(addr);
-	if (unlikely(sz >= 0 && sz < bytes)) {
+	if (unlikely(sz >= 0 && sz < (int)bytes)) {
 		if (!__builtin_constant_p(bytes))
 			copy_overflow(sz, bytes);
 		else if (is_source)
